tp-itba-foundations-2/etl/requirements.txt
pandas
numpy
psycopg2-binary
requests
xlrd